Problématique avec la classe Form.php. Il n'est pas forcément possible d'appeler des fonctions dans les classes Article et Comment depuis Form. Les fonctions ont été réinsérées dans les classes Article et Comment.
