+++ controllers/Backend.php
@@ -26,7 +26,7 @@ class Backend
         $model_article = new Article();
         $model_comment = new Comment();
         $model_user = new User();
-        $model_contact = new Form();
+        $model_form= new Form();
         
         $page="backend/write";
         $topbar="backend/topbar";
@@ -38,7 +38,7 @@ class Backend
         $model_article = new Article();
         $model_comment = new Comment();
         $model_user = new User();
-        $model_contact = new Form();
+        $model_form = new Form();
         
         $page="backend/settings";
         $topbar="backend/topbar";
@@ -72,7 +72,7 @@ class Backend
         $model_article = new Article();
         $model_comment = new Comment();
         $model_user = new User();
-        $model_contact = new Form();
+        $model_form = new Form();
         
         $page="backend/password";
         $topbar="backend/topbar";
@@ -84,7 +84,7 @@ class Backend
         $model_article = new Article();
         $model_comment = new Comment();
         $model_user = new User();
-        $model_contact = new Form();
+        $model_form = new Form();
         
         $page="backend/postback";
         $topbar="backend/topbar";
